Now export all OBJs at the origin and apply Blender's transform to them
This will let us do instancing and animation much easier since otherwise
Blender would export our initial or root instance as being where the
object is in the world. So in that case all our transforms must be
relative to some original mesh location, which would be a real pain.