Add support for instancing. It's a bit hacky, but best we can do i think
We expect the original mesh to be named something like "Cube" and
instances of it to be named "Cube.001", "Cube.002" etc. So look for
objects with names like "Cube.\d+" and make them reference an object
named "Cube" instead. We also remove these instances from the selection
so they don't get exported to the OBJ file.